+++ src/helper/helper.ts
@@ -12,13 +12,3 @@ export function getPos(deltaX = 0, deltaY = 0, deltaWidth = 0, deltaHeight = 0) 
   return { x, y, width, height };
 }
 
-export function drawRect(width: number, height: number) {
-  const pos = getPos(0, 0, width, height);
-  robot.setMouseDelay(200);
-  robot.mouseToggle('down');
-  robot.moveMouse(pos.width, pos.y);
-  robot.moveMouse(pos.width, pos.height);
-  robot.moveMouse(pos.x, pos.height);
-  robot.moveMouse(pos.x, pos.y);
-  robot.mouseToggle('up');
-}
